# PlaceCar_2024_Angular/public/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "PlaceCar API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Factures/Client/{clientId}/NonPaye": {
      "get": {
        "tags": [
          "Facture"
        ],
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReadFactureDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReadFactureDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReadFactureDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Factures/Client/{clientId}/Paye": {
      "get": {
        "tags": [
          "Facture"
        ],
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReadFactureDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReadFactureDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReadFactureDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Factures/Statistics/parking/{parkingId}": {
      "get": {
        "tags": [
          "Facture"
        ],
        "parameters": [
          {
            "name": "parkingId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StatFacturesBO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatFacturesBO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatFacturesBO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Parkings": {
      "post": {
        "tags": [
          "Parking"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/parkingDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/parkingDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/parkingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Parkings/Pays/{paysId}/Ville/{nomVille}": {
      "get": {
        "tags": [
          "Parking"
        ],
        "parameters": [
          {
            "name": "paysId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "nomVille",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Parkings/Pays/{paysId}": {
      "get": {
        "tags": [
          "Parking"
        ],
        "parameters": [
          {
            "name": "paysId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Parkings/Villes/Pays/{paysId}": {
      "get": {
        "tags": [
          "Parking"
        ],
        "parameters": [
          {
            "name": "paysId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Parkings/Employee/{employeeId}": {
      "get": {
        "tags": [
          "Parking"
        ],
        "parameters": [
          {
            "name": "employeeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Parkings/id": {
      "get": {
        "tags": [
          "Parking"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReadParkingDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadParkingDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadParkingDTO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ReadFactureDTO": {
        "type": "object",
        "properties": {
          "facT_Id": {
            "type": "integer",
            "format": "int32"
          },
          "facT_Somme": {
            "type": "number",
            "format": "double"
          },
          "facT_Date": {
            "type": "string",
            "format": "date-time"
          },
          "reS_Id": {
            "type": "integer",
            "format": "int32"
          },
          "reS_DateReservation": {
            "type": "string",
            "format": "date-time"
          },
          "reS_DateDebut": {
            "type": "string",
            "format": "date-time"
          },
          "reS_DateFin": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ReadParkingDTO": {
        "type": "object",
        "properties": {
          "parK_Id": {
            "type": "integer",
            "format": "int32"
          },
          "parK_Nom": {
            "type": "string",
            "nullable": true
          },
          "parK_NbEtages": {
            "type": "integer",
            "format": "int32"
          },
          "parK_NbPlaces": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "StatFacturesBO": {
        "type": "object",
        "properties": {
          "factNonPaye": {
            "type": "integer",
            "format": "int32"
          },
          "factPayee": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "parkingDto": {
        "type": "object",
        "properties": {
          "parK_Nom": {
            "type": "string",
            "nullable": true
          },
          "parK_NbEtages": {
            "type": "integer",
            "format": "int32"
          },
          "parK_NbPlaces": {
            "type": "integer",
            "format": "int32"
          },
          "adrS_Numero": {
            "type": "integer",
            "format": "int32"
          },
          "adrS_NomRue": {
            "type": "string",
            "nullable": true
          },
          "adrS_Ville": {
            "type": "string",
            "nullable": true
          },
          "adrS_Latitude": {
            "type": "number",
            "format": "double"
          },
          "adrS_Longitude": {
            "type": "number",
            "format": "double"
          },
          "payS_Id": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Encoder votre toke; Exemple : Bearer {token}",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}